Import formatDescription util from utils file
