+++ packages/manager/src/features/Support/SupportTickets/SupportTicketDialog.tsx
@@ -21,14 +21,12 @@ import {
   SCHEMA_MAP,
   SEVERITY_LABEL_MAP,
   SEVERITY_OPTIONS,
-  SMTP_FIELD_NAME_TO_LABEL_MAP,
   TICKET_SEVERITY_TOOLTIP_TEXT,
   TICKET_TYPE_MAP,
-  TICKET_TYPE_TO_CUSTOM_FIELD_KEYS_MAP,
 } from './constants';
 import { SupportTicketProductSelectionFields } from './SupportTicketProductSelectionFields';
 import { SupportTicketSMTPFields } from './SupportTicketSMTPFields';
-import { useTicketSeverityCapability } from './ticketUtils';
+import { formatDescription, useTicketSeverityCapability } from './ticketUtils';
 
 import type { FileAttachment } from '../index';
 import type { AttachmentError } from '../SupportTicketDetail/SupportTicketDetail';
@@ -129,44 +127,6 @@ export const getInitialValue = (
   return fromProps ?? fromStorage ?? '';
 };
 
-/**
- * formatDescription
- *
- * When variant ticketTypes include additional fields, fields must concat to one description string.
- * For readability, replace field names with field labels and format the description in Markdown.
- * @param values - the form payload, which can either be the general fields, or the general fields plus any custom fields
- * @param ticketType - either 'general' or a custom ticket type (e.g. 'smtp')
- *
- * @returns a description string containing custom fields in Markdown format
- */
-export const formatDescription = (
-  values: AllSupportTicketFormFields | SupportTicketFormFields,
-  ticketType: TicketType
-) => {
-  type customFieldTuple = [string, string | undefined];
-  const customFields: customFieldTuple[] = Object.entries(
-    values
-  ).filter(([key, _value]: customFieldTuple) =>
-    TICKET_TYPE_TO_CUSTOM_FIELD_KEYS_MAP[ticketType]?.includes(key)
-  );
-
-  // If there are no custom fields, just return the initial description.
-  if (customFields.length === 0) {
-    return values.description;
-  }
-
-  // Add all custom fields to the description in the ticket body, to be viewed on ticket details page and by Customer Support.
-  return customFields
-    .map(([key, value]) => {
-      let label = key;
-      if (ticketType === 'smtp') {
-        label = SMTP_FIELD_NAME_TO_LABEL_MAP[key];
-      }
-      return `**${label}**\n${value ? value : 'No response'}`;
-    })
-    .join('\n\n');
-};
-
 export const SupportTicketDialog = (props: SupportTicketDialogProps) => {
   const {
     open,
@@ -262,7 +222,7 @@ export const SupportTicketDialog = (props: SupportTicketDialogProps) => {
   const onSubmit = form.handleSubmit(async (values) => {
     // console.log(form.getValues());
     // const { onSuccess } = props;
-    const _description = formatDescription(values, ticketType);
+    form.setValue('description', formatDescription(values, ticketType));
 
     if (!['general', 'none'].includes(entityType) && !entityId) {
       form.setError('entityId', {
@@ -273,7 +233,7 @@ export const SupportTicketDialog = (props: SupportTicketDialogProps) => {
     }
     setSubmitting(true);
     createSupportTicket({
-      description: _description,
+      description,
       [entityType]: Number(entityId),
       severity: selectedSeverity,
       summary,
